Reord `zeus.device.gpu`
